+++ changelog.md
@@ -1,4 +1,5 @@
 # v0.1.15
+- Sanitize titles while they are being typed
 - Fix content being displayed when blank
 
 # v0.1.14